feat(components, app): move wizard header and stepmeter to component for pd
move wizard header and stepmeter to component for pd

close AUTH-
