+++ app/src/organisms/Desktop/CalibrateTipLength/AskForCalibrationBlockModal.tsx
@@ -17,11 +17,11 @@ import {
   SecondaryButton,
   SPACING,
   TYPOGRAPHY,
+  WizardHeader,
 } from '@opentrons/components'
 
 import { getTopPortalEl } from '/app/App/portal'
 import { labwareImages } from '/app/local-resources/labware'
-import { WizardHeader } from '/app/molecules/WizardHeader'
 import { setUseTrashSurfaceForTipCal } from '/app/redux/calibration'
 
 import styles from './styles.module.css'
@@ -95,6 +95,7 @@ export function AskForCalibrationBlockModal(props: Props): JSX.Element {
           </Flex>
           <Flex flex="1" justifyContent={JUSTIFY_CENTER}>
             <img
+              alt="Calibration Block"
               className={styles.block_image}
               src={labwareImages[CAL_BLOCK_LOAD_NAME]}
             />
