+++ app/src/organisms/LabwarePositionCheck/__tests__/LPCContentContainer.test.tsx
@@ -3,7 +3,6 @@ import { screen } from '@testing-library/react'
 import { afterEach, beforeEach, describe, it, vi } from 'vitest'
 
 import { renderWithProviders } from '/app/__testing-utils__'
-import { StepMeter } from '/app/atoms/StepMeter'
 import { i18n } from '/app/i18n'
 import { mockLPCContentProps } from '/app/organisms/LabwarePositionCheck/__fixtures__/mockLPCContentProps'
 import { LPCContentContainer } from '/app/organisms/LabwarePositionCheck/LPCContentContainer'
@@ -11,6 +10,7 @@ import { LPCContentContainer } from '/app/organisms/LabwarePositionCheck/LPCCont
 import { ChildNavigation } from '/app/organisms/ODD/ChildNavigation'
 
 import type { ComponentProps } from 'react'
+import type { StepMeter } from '@opentrons/components'
 
 vi.mock('react-redux', async importOriginal => {
   const actual = await importOriginal<typeof useSelector>()
@@ -20,7 +20,13 @@ vi.mock('react-redux', async importOriginal => {
   }
 })
 vi.mock('/app/organisms/ODD/ChildNavigation')
-vi.mock('/app/atoms/StepMeter')
+vi.mock('@opentrons/components', async importOriginal => {
+  const actualComponents = await importOriginal<typeof StepMeter>()
+  return {
+    ...actualComponents,
+    StepMeter: vi.fn(({ children }) => <div>MOCK_STEP_METER</div>),
+  }
+})
 
 const render = (props: ComponentProps<typeof LPCContentContainer>) => {
   return renderWithProviders(<LPCContentContainer {...props} />, {
@@ -42,7 +48,6 @@ describe('LPCContentContainer', () => {
     }
 
     vi.mocked(ChildNavigation).mockReturnValue(<div>MOCK_CHILD_NAV</div>)
-    vi.mocked(StepMeter).mockReturnValue(<div>MOCK_STEP_METER</div>)
     vi.mocked(useSelector).mockReturnValue({
       currentStepIndex: 1,
       totalStepCount: 5,
