+++ app/src/organisms/ODD/NameRobot/ConfirmRobotName.tsx
@@ -9,12 +9,12 @@ import {
   JUSTIFY_CENTER,
   LegacyStyledText,
   SPACING,
+  StepMeter,
   TYPOGRAPHY,
 } from '@opentrons/components'
 
 import screenImage from '/app/assets/images/on-device-display/odd_abstract@x2.png'
 import { MediumButton } from '/app/atoms/buttons'
-import { StepMeter } from '/app/atoms/StepMeter'
 
 const IMAGE_ALT = 'finish setting up a robot'
