+++ app/src/pages/Desktop/Devices/CalibrationDashboard/hooks/useDashboardCalibratePipOffset.tsx
@@ -3,10 +3,9 @@ import { createPortal } from 'react-dom'
 import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
 
-import { ModalShell } from '@opentrons/components'
+import { ModalShell, WizardHeader } from '@opentrons/components'
 
 import { getTopPortalEl } from '/app/App/portal'
-import { WizardHeader } from '/app/molecules/WizardHeader'
 import { CalibratePipetteOffset } from '/app/organisms/Desktop/CalibratePipetteOffset'
 import { LoadingState } from '/app/organisms/Desktop/CalibrationPanels'
 import { pipetteOffsetCalibrationStarted } from '/app/redux/analytics'
