+++ app/src/pages/ODD/ConnectViaWifi/index.tsx
@@ -2,9 +2,13 @@ import { useState } from 'react'
 import { useSelector } from 'react-redux'
 import last from 'lodash/last'
 
-import { DIRECTION_COLUMN, Flex, SPACING } from '@opentrons/components'
+import {
+  DIRECTION_COLUMN,
+  Flex,
+  SPACING,
+  StepMeter,
+} from '@opentrons/components'
 
-import { StepMeter } from '/app/atoms/StepMeter'
 import { DisplayWifiList } from '/app/organisms/ODD/NetworkSettings'
 import { getLocalRobot } from '/app/redux/discovery'
 import * as Networking from '/app/redux/networking'
