+++ components/src/molecules/index.ts
@@ -10,3 +10,4 @@ export * from './ParametersTable'
 export * from './ReleaseNotes'
 export * from './Tabs'
 export * from './TextListTableContent'
+export * from './WizardHeader'
